requirements.txt
fastapi
uvicorn
groq
python-dotenv
twilio
pandas